refactor(api): PR-6 commit 1 — 5 个小 router 抽出 (root/samples/logs/data_exports/tagger)
server.py 4068 → 3944 行（−124），5 个 1-route router 搬到 api/routers/，
`_thumb_response` 共享 helper 升级到 api/responses.py。

  api/routers/root.py           1 route   GET /            302 → /studio/ SPA
  api/routers/samples.py        1 route   GET /samples/{filename}
  api/routers/logs.py           1 route   GET /api/logs/{task_id}
  api/routers/data_exports.py   1 route   GET /api/data-exports
  api/routers/tagger.py         1 route   GET /api/tagger/{name}/check

  api/responses.py + _thumb_response（PR-5 抽 EMPTY_STATE 时单独的，现在
                    samples router + server.py 内 project_thumb / version_thumb
                    都共用同一函数）

server.py 顶部清理：
  - 删 `OUTPUT_DIR` 直接用法但保留 import（test_studio_server fixture
    `monkeypatch.setattr(server, "OUTPUT_DIR", X)` 兼容；samples router 内
    真位置 fixture 也同步 patch）
  - 删 `JSONResponse / RedirectResponse / get_tagger` 死 import
  - `from .api.responses import EMPTY_STATE, _thumb_response`（project_thumb /
    version_thumb 内 3 处 callsite 不动）

测试跟进：tests/test_studio_server.py `isolated_paths` fixture 加 patch
`api.routers.samples.OUTPUT_DIR` + `api.routers.root.WEB_DIST` 同步搬迁
（PR-5 PR description 写过的 lesson — fixture monkeypatch 必须跟 router
搬迁加新位置；现 codify 进 fixture 注释）。

回归：PR-1 安全网 152 + studio_server 44 + datasets 15 + presets/import smoke
全 = **232/232 全绿**。route snapshot 不变（160 routes）。
